Changed mean coverage and max CN data type to float

+++ SVclone/SVprocess/annotate.py
@@ -509,8 +509,8 @@ def preproc_svs(args):
     if len(cfg_file)==0:
         raise ValueError('No configuration file found')
 
-    max_cn       = int(Config.get('BamParameters', 'max_cn'))
-    mean_cov     = int(Config.get('BamParameters', 'mean_cov'))
+    max_cn       = float(Config.get('BamParameters', 'max_cn'))
+    mean_cov     = float(Config.get('BamParameters', 'mean_cov'))
     rlen         = int(Config.get('BamParameters', 'read_len'))
     insert_mean  = int(Config.get('BamParameters', 'insert_mean'))
     insert_std   = int(Config.get('BamParameters', 'insert_std'))
